+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.11.1] - 2023-05-02
+
+### Bug Fixes
+- Missing quorum params (#626)
+
+
 ## [0.11.0] - 2023-04-04
 
 ### Bug Fixes
@@ -3789,7 +3795,7 @@
 - Only run when applicable (#4752)
 - Check git diff on each job (#4770)
 - Checkout code before git diff check (#4779)
-- Add paths 
+- Add paths
 - Bump the timeout for test_coverage (#4864)
 - Migrate localnet to github actions (#4878)
 - Add timeouts (#4912)
@@ -5914,7 +5920,7 @@
 ### Swagger
 
 - Update swagger port (#4498)
-- Remove duplicate blockID 
+- Remove duplicate blockID
 - Define version (#4952)
 - Update (#5257)